+++ src/index.js
@@ -35,7 +35,7 @@ export { nullable } from './nullable';
 export { integer, number, positiveInteger, positiveNumber } from './number';
 export { field, object, pojo } from './object';
 export { optional } from './optional';
-export { regex, string } from './string';
+export { email, regex, string, url } from './string';
 export { tuple2 } from './tuple';
 
 export type { Guard, Decoder };